Выносит вспомогательные функции из main
